Initial commit of 'C' version for testing (see the help).
The application is designed to create simple WAV files for testing audio
functionality, and should save a lot of time over using Audacity, etc.

Testing of this version:

 * Simple tests of each waveform type only - not exhaustive.
 * Running under valgrind.
 * CodeChecker.

There are still some TODOs but it serves my current purpose as-is.

Initial commit of 'C' version with basic functionality (see the help).

The application is designed to create simple WAV files for testing audio
functionality, and should save a lot of time over using Audacity, etc.

Testing of this version:

 * Simple tests of each waveform type only - not exhaustive.
 * Running under valgrind.
 * CodeChecker.

There are still some TODOs but it serves my current purpose as-is.