Fix truncated duration due to integer overflow.
The num_samples calculation used to control the generation loops could
overflow if the duration was specified in terms of more than a few tens
of seconds.

The Some small units such as bit-depth have had their storage size
reduced, but this is not an integral part of the duration fix.